requirements.txt
psycopg2
aiofiles
jinja2
python-multipart
starlette
python-jose
passlib
cffi
bcrypt
